+++ dynamorm/types/base.py
@@ -36,6 +36,6 @@ class DynamORMSchema(object):
         )
 
     @staticmethod
-    def base_field_type():
-        """Returns the class that all fields in the schema will inherit from"""
-        raise NotImplementedError("Child class must implement base_field_type")
+    def base_schema_type():
+        """Returns the base class used for schemas of this type"""
+        raise NotImplementedError("Child class must implement base_schema_type")